+++ backend/api/tests/views/test_views_newsfeed.py
@@ -44,7 +44,6 @@ class NewsFeedViewTests(APITestCase):
         canceled_cp = factory.create_computeplan(status=ComputePlan.Status.PLAN_STATUS_CANCELED)
         failed_cp = factory.create_computeplan(status=ComputePlan.Status.PLAN_STATUS_FAILED)
         failed_cp.failed_task_key = str(uuid4())
-        failed_cp.failed_task_category = "TASK_UNKNOWN"
         failed_cp.save()
         done_cp = factory.create_computeplan(status=ComputePlan.Status.PLAN_STATUS_DONE)
         datamanager = factory.create_datamanager()
@@ -69,7 +68,6 @@ class NewsFeedViewTests(APITestCase):
                 "timestamp": failed_cp.end_date.isoformat().replace("+00:00", "Z"),
                 "detail": {
                     "first_failed_task_key": failed_cp.failed_task_key,
-                    "task_category": "TASK_UNKNOWN",
                 },
             },
             # PLAN_STATUS_CANCELED