+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Test only field for data samples (#551)
+- BREAKING: failed task category from compute plan API response([#525](https://github.com/Substra/substra-backend/pull/525))
 
 ## [0.34.1](https://github.com/Substra/substra-backend/releases/tag/0.34.1) 2023-01-09